approximate a login, except the login isn't actually stored on the server
you don't want to know what ungodly amount of time i couldn't figure out why sign in with oauth wasn't working... because i was running it on the server. i just assumed if you run email/pass auth there, you can also run oauth there. no.

i had to dig into the auth source code to realize, hey it doesn't redirect you if you're not in the browser... that's because it's a server. obviously. you have to feed the url to client